+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 BUG FIXES:
 
 * Resource: `tencentcloud_instance` fix `allocate_public_ip` inconsistency when eip is attached to the cvm.
+* Resource: `tencentcloud_mysql_instance` fix auto-forcenew on `charge_type` and `pay_type` when upgrading terraform version. ([#459](https://github.com/terraform-providers/terraform-provider-tencentcloud/pull/459)).
 
 ## 1.38.1 (June 30, 2020)